fix local execution
