--- canisters/chain_fusion/src/job/calculate_result.rs
@@ -1,9 +1,0 @@
-pub fn fibonacci(n: u64) -> u64 {
-    if n == 0 {
-        0
-    } else if n == 1 {
-        1
-    } else {
-        fibonacci(n - 1) + fibonacci(n - 2)
-    }
-}
