+++ canisters/chain_fusion/src/job/submit_result.rs
@@ -1,76 +1,75 @@
-use ethers_core::{types::U256, utils::keccak256};
-use ic_cdk::println;
+use ethers_core::{abi::Token, types::U256};
+use evm_rpc_canister_types::{SendRawTransactionStatus, EVM_RPC};
+use ic_evm_utils::eth_send_raw_transaction::{contract_interaction, ContractDetails};
 
-use crate::{
-    eth_send_raw_transaction::{create_sign_request, send_raw_transaction},
-    evm_rpc::SendRawTransactionStatus,
-    evm_signer, fees,
-    state::{mutate_state, read_state},
-};
-use ethers_core::abi::AbiEncode;
+use crate::state::{mutate_state, read_state, State};
 
-// todo: properly name this function
 pub async fn submit_result(job_id: U256) {
-    println! {"Submitting result for job_id: {job_id}"};
-    let function_signature = "executeJob(uint256)";
+    // get necessary global state
+    let contract_address = &read_state(State::get_logs_addresses)[0];
+    let rpc_services = read_state(State::rpc_services);
+    let nonce = read_state(State::nonce);
+    let key_id = read_state(State::key_id);
 
-    let selector = &keccak256(function_signature.as_bytes())[0..4];
-    let args = (job_id).encode();
-    let mut data = Vec::from(selector);
-    data.extend(args);
+    let abi_json = r#"
+   [
+        {
+            "type": "function",
+            "name": "executeJob",
+            "inputs": [
+                {
+                    "name": "_job_id",
+                    "type": "uint256",
+                    "internalType": "uint256"
+                }
+            ],
+            "outputs": [],
+            "stateMutability": "nonpayable"
+        }
+   ]
+   "#;
 
-    let fee_estimates = fees::estimate_transaction_fees(9).await;
+    let abi =
+        serde_json::from_str::<ethers_core::abi::Contract>(abi_json).expect("should serialise");
 
-    //let gas_limit = U256::from(1000000000);
-    // let manual_fee_estimates = fees::FeeEstimates {
-    //     max_fee_per_gas: gas_limit,
-    //     max_priority_fee_per_gas: U256::from(10000000),
-    // };
+    let contract_details = ContractDetails {
+        contract_address: contract_address.clone(),
+        abi: &abi,
+        function_name: "executeJob",
+        args: &[Token::Uint(job_id)],
+    };
 
-    // todo: replace this manual estimation of the gas consumption with a better one
-    let manual_gas_limit = U256::from(1000000);
+    // set the gas
+    let gas = Some(U256::from(1000000));
 
-    let contract_address = read_state(|s| s.get_logs_address[0].clone());
-
-    let request = create_sign_request(
-        U256::from(0),
-        Some(contract_address),
-        None,
-        manual_gas_limit,
-        Some(data),
-        fee_estimates,
-        //manual_fee_estimates,
+    // interact with the contract, this calls `eth_sendRawTransaction` under the hood
+    let status = contract_interaction(
+        contract_details,
+        gas,
+        rpc_services,
+        nonce,
+        key_id,
+        vec![],
+        EVM_RPC,
     )
     .await;
 
-    println! {"Signing request "};
-
-    let tx = evm_signer::sign_transaction(request).await;
-
-    println! {"Submitting signed request  "};
-
-    let status = send_raw_transaction(tx.clone()).await;
-
-    println!("Transaction sent: {:?}", tx);
-
+    // if the transaction
     match status {
         SendRawTransactionStatus::Ok(transaction_hash) => {
-            println!("Success {transaction_hash:?}");
+            ic_cdk::println!("Success {transaction_hash:?}");
             mutate_state(|s| {
                 s.nonce += U256::from(1);
             });
         }
         SendRawTransactionStatus::NonceTooLow => {
-            println!("Nonce too low");
-            // print current nonce from state
-            let nonce = read_state(|s| s.nonce.clone());
-            println!("Current nonce: {nonce}");
+            ic_cdk::println!("Nonce too low");
         }
         SendRawTransactionStatus::NonceTooHigh => {
-            println!("Nonce too high");
+            ic_cdk::println!("Nonce too high");
         }
         SendRawTransactionStatus::InsufficientFunds => {
-            println!("Insufficient funds");
+            ic_cdk::println!("Insufficient funds");
         }
     }
 }
